requirements.txt
psycopg2 
python-dotenv
ollama
streamlit